[software] usb: Fix image cutoff issue
It turns out that it is because when doubling the resolution, I forgot to bump the packet format to uint16_t.

+++ Software/src/drivers/laser_canvas.h
@@ -7,8 +7,8 @@
 class laser_canvas {
  public:
   struct tuple {
-    std::uint8_t x;
-    std::uint8_t y;
+    std::uint16_t x;
+    std::uint16_t y;
     bool state;
   };
